+++ RELEASE_NOTES.md
@@ -2,9 +2,13 @@
 
 If you like this project, please [gift a ★★★★★ rating](https://marketplace.visualstudio.com/items?itemName=lsoft.FreeAIr).
 
+Please report any bugs to the [github repo](https://github.com/lsoft/FreeAIr).
+
+If you are enjoying FreeAIr to the enough level to donate, there are a a [lot of small cancer patients](https://advitausa.org/au/index.php/donate/) that need your help.
+
 ## 0.2.3
 
-- Added a natural language search across whole solution.
+- Added a natural language search (across whole solution or current project).
 - Implemented UI improvements.
 - Added a github MCP server and Visual Studio MCP server and its tool.
 - And a lot more.